migrate HealthCache
