+++ programs/marginfi/src/instructions/marginfi_account/borrow.rs
@@ -6,7 +6,6 @@ use crate::{
     prelude::{MarginfiError, MarginfiGroup, MarginfiResult},
     state::{
         bank::{BankImpl, BankVaultType},
-        health_cache::HealthCache,
         marginfi_account::{BankAccountWrapper, MarginfiAccount, RiskEngine, ACCOUNT_DISABLED},
     },
     utils::{self, validate_asset_tags},
@@ -16,7 +15,7 @@ use anchor_lang::solana_program::{clock::Clock, sysvar::Sysvar};
 use anchor_spl::token_interface::{TokenAccount, TokenInterface};
 use bytemuck::Zeroable;
 use fixed::types::I80F48;
-use marginfi_type_crate::types::Bank;
+use marginfi_type_crate::types::{Bank, HealthCache};
 
 /// 1. Accrue interest
 /// 2. Create the user's bank account for the asset borrowed if it does not exist yet
