+++ programs/marginfi/src/instructions/marginfi_account/withdraw.rs
@@ -5,7 +5,6 @@ use crate::{
     prelude::*,
     state::{
         bank::{BankImpl, BankVaultType},
-        health_cache::HealthCache,
         marginfi_account::{BankAccountWrapper, MarginfiAccount, RiskEngine, ACCOUNT_DISABLED},
     },
     utils,
@@ -15,7 +14,7 @@ use anchor_lang::solana_program::{clock::Clock, sysvar::Sysvar};
 use anchor_spl::token_interface::{TokenAccount, TokenInterface};
 use bytemuck::Zeroable;
 use fixed::types::I80F48;
-use marginfi_type_crate::types::Bank;
+use marginfi_type_crate::types::{Bank, HealthCache};
 
 /// 1. Accrue interest
 /// 2. Find the user's existing bank account for the asset withdrawn
