+++ programs/marginfi/src/instructions/marginfi_group/handle_bankruptcy.rs
@@ -10,7 +10,6 @@ use crate::{
     prelude::MarginfiError,
     state::{
         bank::{BankImpl, BankVaultType},
-        health_cache::HealthCache,
         marginfi_account::{BankAccountWrapper, MarginfiAccount, RiskEngine, ACCOUNT_DISABLED},
         marginfi_group::MarginfiGroup,
     },
@@ -20,7 +19,7 @@ use anchor_lang::prelude::*;
 use anchor_spl::token_interface::{TokenAccount, TokenInterface};
 use bytemuck::Zeroable;
 use fixed::types::I80F48;
-use marginfi_type_crate::types::Bank;
+use marginfi_type_crate::types::{Bank, HealthCache};
 use std::cmp::{max, min};
 
 /// Handle a bankrupt marginfi account.
