+++ programs/marginfi/src/state/mod.rs
@@ -1,7 +1,6 @@
 pub mod bank;
 pub mod bank_cache;
 pub mod emode;
-pub mod health_cache;
 pub mod marginfi_account;
 pub mod marginfi_group;
 pub mod price;
