+++ type-crate/src/types/health_cache.rs
@@ -1,13 +1,24 @@
 use bytemuck::{Pod, Zeroable};
 
+use crate::{assert_struct_align, assert_struct_size};
+
 use super::{WrappedI80F48, MAX_LENDING_ACCOUNT_BALANCES};
+
+#[cfg(feature = "anchor")]
+use {anchor_lang::prelude::*, type_layout::TypeLayout};
 
 pub const HEALTHY: u32 = 1;
 pub const ENGINE_OK: u32 = 2;
 pub const ORACLE_OK: u32 = 4;
 
+assert_struct_size!(HealthCache, 304);
+assert_struct_align!(HealthCache, 8);
 #[repr(C)]
-#[derive(Debug, PartialEq, Pod, Zeroable, Copy, Clone)]
+#[cfg_attr(
+    feature = "anchor",
+    derive(AnchorDeserialize, AnchorSerialize, TypeLayout)
+)]
+#[derive(Debug, PartialEq, Eq, Pod, Zeroable, Copy, Clone)]
 /// A read-only cache of the internal risk engine's information. Only valid in borrow/withdraw if
 /// the tx does not fail. To see the state in any context, e.g. to figure out if the risk engine is
 /// failing due to some bad price information, use `pulse_health`.
@@ -111,6 +122,19 @@ impl HealthCache {
             self.flags |= ENGINE_OK;
         } else {
             self.flags &= !ENGINE_OK;
+        }
+    }
+
+    /// True if the engine did not detect an oracle input issue in the last health pulse
+    pub fn is_oracle_ok(&self) -> bool {
+        self.flags & ORACLE_OK != 0
+    }
+
+    pub fn set_oracle_ok(&mut self, ok: bool) {
+        if ok {
+            self.flags |= ORACLE_OK;
+        } else {
+            self.flags &= !ORACLE_OK;
         }
     }
 }
